+++ newrelic-security-agent/src/main/java/com/newrelic/agent/security/instrumentator/httpclient/RestClient.java
@@ -2,6 +2,8 @@ package com.newrelic.agent.security.instrumentator.httpclient;
 
 import com.newrelic.agent.security.AgentInfo;
 import com.newrelic.agent.security.intcodeagent.filelogging.FileLoggerThreadPool;
+import com.newrelic.agent.security.intcodeagent.logging.IAgentConstants;
+import com.newrelic.agent.security.intcodeagent.models.FuzzRequestBean;
 import com.newrelic.api.agent.security.utils.logging.LogLevel;
 import com.newrelic.agent.security.intcodeagent.models.javaagent.FuzzFailEvent;
 import com.newrelic.agent.security.intcodeagent.websocket.EventSendPool;
@@ -15,6 +17,7 @@ import java.io.IOException;
 import java.io.InterruptedIOException;
 import java.security.cert.CertificateException;
 import java.util.HashSet;
+import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 public class RestClient {
@@ -118,7 +121,45 @@ public class RestClient {
         return clientThreadLocal.get();
     }
 
-    public void fireRequest(Request request, int repeatCount, String fuzzRequestId) {
+    public void fireRequest(FuzzRequestBean httpRequest, List<String> endpoints, int repeatCount, String fuzzRequestId){
+
+        int responseCode = 999;
+        if(endpoints.isEmpty()){
+            Request request = RequestUtils.generateK2Request(httpRequest, String.format(IAgentConstants.ENDPOINT_LOCALHOST_S, httpRequest.getProtocol(), httpRequest.getServerPort()));
+            if (request != null) {
+                try {
+                    responseCode = RestClient.getInstance().fireRequest(request, repeatCount + endpoints.size() -1, fuzzRequestId);
+                } catch (SSLException e) {
+                    logger.log(LogLevel.FINER, String.format(CALL_FAILED_REQUEST_S_REASON, request), e, RestClient.class.getName());
+                    logger.postLogMessageIfNecessary(LogLevel.WARNING,
+                            String.format(CALL_FAILED_REQUEST_S_REASON, fuzzRequestId),
+                            e, RestRequestProcessor.class.getName());
+                    RestRequestThreadPool.getInstance().getProcessedIds().putIfAbsent(fuzzRequestId, new HashSet<>());
+                    // TODO: Add to fuzz fail count in HC and remove FuzzFailEvent if not needed.
+                    FuzzFailEvent fuzzFailEvent = new FuzzFailEvent(AgentInfo.getInstance().getApplicationUUID());
+                    fuzzFailEvent.setFuzzHeader(request.header(ServletHelper.CSEC_IAST_FUZZ_REQUEST_ID));
+                    EventSendPool.getInstance().sendEvent(fuzzFailEvent);
+                }
+            }
+            return;
+        }
+        for (String endpoint : endpoints) {
+            Request request = RequestUtils.generateK2Request(httpRequest, endpoint);
+            try {
+                if (request != null) {
+                    responseCode = RestClient.getInstance().fireRequest(request, repeatCount + endpoints.size() -1, fuzzRequestId);
+                }
+                if(responseCode == 301){continue;}
+                break;
+            } catch (SSLException e){
+                logger.log(LogLevel.FINER, String.format(CALL_FAILED_REQUEST_S_REASON, request), e, RestClient.class.getName());
+            }
+        }
+
+
+    }
+
+    public int fireRequest(Request request, int repeatCount, String fuzzRequestId) throws SSLException {
         OkHttpClient client = clientThreadLocal.get();
 
         logger.log(LogLevel.FINER, String.format(FIRING_REQUEST_METHOD_S, request.method()), RestClient.class.getName());
@@ -143,9 +184,13 @@ public class RestClient {
             if (client.connectionPool() != null) {
                 client.connectionPool().evictAll();
             }
+            return response.code();
+        } catch (SSLException e){
+            logger.log(LogLevel.FINE, String.format("Request failed due to SSL Exception %s ", request, e), RestClient.class.getName());
+            throw e;
         } catch (InterruptedIOException e){
             if(repeatCount >= 0){
-                fireRequest(request, --repeatCount, fuzzRequestId);
+                return fireRequest(request, --repeatCount, fuzzRequestId);
             }
         } catch (IOException e) {
             logger.log(LogLevel.FINER, String.format(CALL_FAILED_REQUEST_S_REASON, request), e, RestClient.class.getName());
@@ -159,6 +204,7 @@ public class RestClient {
             EventSendPool.getInstance().sendEvent(fuzzFailEvent);
         }
 
+        return 999;
     }
 
     public boolean isConnected() {