+++ newrelic-security-agent/src/main/java/com/newrelic/agent/security/instrumentator/httpclient/RestRequestProcessor.java
@@ -6,6 +6,7 @@ import com.newrelic.agent.security.AgentInfo;
 import com.newrelic.agent.security.instrumentator.os.OsVariablesInstance;
 import com.newrelic.agent.security.instrumentator.utils.CallbackUtils;
 import com.newrelic.agent.security.intcodeagent.filelogging.FileLoggerThreadPool;
+import com.newrelic.api.agent.security.NewRelicSecurity;
 import com.newrelic.api.agent.security.utils.logging.LogLevel;
 import com.newrelic.agent.security.intcodeagent.models.FuzzRequestBean;
 import com.newrelic.agent.security.intcodeagent.models.javaagent.IntCodeControlCommand;
@@ -16,9 +17,11 @@ import com.newrelic.api.agent.security.instrumentation.helpers.GenericHelper;
 import okhttp3.Request;
 import org.apache.commons.lang3.StringUtils;
 
+import javax.net.ssl.SSLException;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.Callable;
 
 /**
@@ -100,10 +103,8 @@ public class RestRequestProcessor implements Callable<Boolean> {
                 MonitorGrpcFuzzFailRequestQueueThread.submitNewTask();
                 GrpcClientRequestReplayHelper.getInstance().addToRequestQueue(new ControlCommandDto(controlCommand.getId(), httpRequest, payloadList));
             } else {
-                Request request = RequestUtils.generateK2Request(httpRequest);
-                if(request != null) {
-                    RestClient.getInstance().fireRequest(request, repeatCount, controlCommand.getId());
-                }
+                List<String> endpoints = prepareAllEndpoints(NewRelicSecurity.getAgent().getApplicationConnectionConfig());
+                RestClient.getInstance().fireRequest(httpRequest, endpoints, repeatCount + endpoints.size() -1, controlCommand.getId());
             }
             return true;
         } catch (JsonProcessingException e){
@@ -124,6 +125,19 @@ public class RestRequestProcessor implements Callable<Boolean> {
             throw e;
         }
         return true;
+    }
+
+    private List<String> prepareAllEndpoints(Map<Integer, String> applicationConnectionConfig) {
+        List<String> endpoitns = new ArrayList<>();
+        for (Map.Entry<Integer, String> connectionConfig : applicationConnectionConfig.entrySet()) {
+            endpoitns.add(String.format("%s://localhost:%s", connectionConfig.getValue(), connectionConfig.getKey()));
+            endpoitns.add(String.format("%s://localhost:%s", toggleProtocol(connectionConfig.getValue()), connectionConfig.getKey()));
+        }
+        return endpoitns;
+    }
+
+    private String toggleProtocol(String value) {
+        return StringUtils.equalsAnyIgnoreCase(value, "https")? "http": "https";
     }
 
     public static void processControlCommand(IntCodeControlCommand command) {